Split authorizer.rs
Issue-Id: #22
