+++ agent/src/control_interface/authorizer/path.rs
@@ -1,0 +1,38 @@
+// Copyright (c) 2024 Elektrobit Automotive GmbH
+//
+// This program and the accompanying materials are made available under the
+// terms of the Apache License, Version 2.0 which is available at
+// https://www.apache.org/licenses/LICENSE-2.0.
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+// WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+// License for the specific language governing permissions and limitations
+// under the License.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+use common::PATH_SEPARATOR;
+
+#[derive(Clone, Debug)]
+pub struct Path {
+    pub sections: Vec<String>,
+}
+
+impl From<&str> for Path {
+    fn from(value: &str) -> Self {
+        Self {
+            sections: if value.is_empty() {
+                Vec::new()
+            } else {
+                value.split(PATH_SEPARATOR).map(Into::into).collect()
+            },
+        }
+    }
+}
+
+impl ToString for Path {
+    fn to_string(&self) -> String {
+        self.sections.join(".")
+    }
+}
