+++ agent/src/control_interface/authorizer/path_pattern.rs
@@ -1,0 +1,246 @@
+// Copyright (c) 2024 Elektrobit Automotive GmbH
+//
+// This program and the accompanying materials are made available under the
+// terms of the Apache License, Version 2.0 which is available at
+// https://www.apache.org/licenses/LICENSE-2.0.
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+// WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+// License for the specific language governing permissions and limitations
+// under the License.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+use super::path::Path;
+use common::PATH_SEPARATOR;
+
+const WILDCARD_SYMBOL: &str = "*";
+
+pub type PathPatternMatchReason = String;
+pub trait PathPattern {
+    fn matches(&self, other: &Path) -> (bool, PathPatternMatchReason);
+}
+
+impl<T: PathPattern + std::fmt::Debug> PathPattern for Vec<T> {
+    fn matches(&self, path: &Path) -> (bool, PathPatternMatchReason) {
+        for r in self {
+            if let (true, reason) = r.matches(path) {
+                return (true, reason);
+            }
+        }
+        (false, String::new())
+    }
+}
+
+#[derive(Clone, Debug, PartialEq)]
+pub struct AllowPathPattern {
+    sections: Vec<PathPatternSection>,
+}
+
+impl AllowPathPattern {
+    #[cfg(test)]
+    pub fn test_value(name: &str) -> Self {
+        Self {
+            sections: vec![PathPatternSection::String(name.into())],
+        }
+    }
+}
+
+impl From<&str> for AllowPathPattern {
+    fn from(value: &str) -> Self {
+        Self {
+            sections: if value.is_empty() {
+                Vec::new()
+            } else {
+                value.split(PATH_SEPARATOR).map(Into::into).collect()
+            },
+        }
+    }
+}
+
+impl PathPattern for AllowPathPattern {
+    fn matches(&self, other: &Path) -> (bool, PathPatternMatchReason) {
+        if self.sections.len() > other.sections.len() {
+            return (false, String::new());
+        }
+        for (a, b) in self.sections.iter().zip(other.sections.iter()) {
+            if !a.matches(b) {
+                return (false, String::new());
+            }
+        }
+
+        (
+            true,
+            self.sections
+                .iter()
+                .map(ToString::to_string)
+                .collect::<Vec<_>>()
+                .join("."),
+        )
+    }
+}
+
+#[derive(Clone, Debug, PartialEq)]
+pub struct DenyPathPattern {
+    sections: Vec<PathPatternSection>,
+}
+
+impl From<&str> for DenyPathPattern {
+    fn from(value: &str) -> Self {
+        Self {
+            sections: if value.is_empty() {
+                Vec::new()
+            } else {
+                value.split(PATH_SEPARATOR).map(Into::into).collect()
+            },
+        }
+    }
+}
+
+impl PathPattern for DenyPathPattern {
+    fn matches(&self, other: &Path) -> (bool, PathPatternMatchReason) {
+        for (a, b) in self.sections.iter().zip(other.sections.iter()) {
+            if !a.matches(b) {
+                return (false, String::new());
+            }
+        }
+        (
+            true,
+            self.sections
+                .iter()
+                .map(ToString::to_string)
+                .collect::<Vec<_>>()
+                .join("."),
+        )
+    }
+}
+
+#[derive(Clone, Debug, PartialEq)]
+enum PathPatternSection {
+    Wildcard,
+    String(String),
+}
+
+impl From<&str> for PathPatternSection {
+    fn from(value: &str) -> Self {
+        if value == WILDCARD_SYMBOL {
+            Self::Wildcard
+        } else {
+            Self::String(value.into())
+        }
+    }
+}
+
+impl ToString for PathPatternSection {
+    fn to_string(&self) -> String {
+        match self {
+            PathPatternSection::Wildcard => "*".into(),
+            PathPatternSection::String(s) => s.clone(),
+        }
+    }
+}
+
+impl PathPatternSection {
+    pub fn matches(&self, other: &String) -> bool {
+        match self {
+            PathPatternSection::Wildcard => true,
+            PathPatternSection::String(self_string) => self_string == other,
+        }
+    }
+}
+
+//////////////////////////////////////////////////////////////////////////////
+//                 ########  #######    #########  #########                //
+//                    ##     ##        ##             ##                    //
+//                    ##     #####     #########      ##                    //
+//                    ##     ##                ##     ##                    //
+//                    ##     #######   #########      ##                    //
+//////////////////////////////////////////////////////////////////////////////
+
+#[cfg(test)]
+mod tests {
+    use crate::control_interface::authorizer::{AllowPathPattern, DenyPathPattern, PathPattern};
+
+    #[test]
+    fn utest_allow_path_pattern() {
+        let p = AllowPathPattern::from("some.pre.fix");
+
+        assert!(p.matches(&"some.pre.fix".into()).0);
+        assert!(p.matches(&"some.pre.fix.test".into()).0);
+        assert!(!p.matches(&"".into()).0);
+        assert!(!p.matches(&"some.pre".into()).0);
+        assert!(!p.matches(&"some.pre.fixtest".into()).0);
+        assert!(!p.matches(&"some.pre.test".into()).0);
+        assert!(!p.matches(&"some.pre.test.2".into()).0);
+    }
+
+    #[test]
+    fn utest_allow_path_pattern_with_wildcard() {
+        let p = AllowPathPattern::from("some.*.fix");
+
+        assert!(p.matches(&"some.pre.fix".into()).0);
+        assert!(p.matches(&"some.pre.fix.test".into()).0);
+        assert!(!p.matches(&"some.pre".into()).0);
+        assert!(!p.matches(&"some.pre.fixtest".into()).0);
+        assert!(!p.matches(&"some.pre.test".into()).0);
+        assert!(!p.matches(&"some.pre.test.2".into()).0);
+        assert!(p.matches(&"some.pre2.fix".into()).0);
+        assert!(p.matches(&"some.pre2.fix.test".into()).0);
+        assert!(!p.matches(&"some.pre2".into()).0);
+        assert!(!p.matches(&"some.pre2.fixtest".into()).0);
+        assert!(!p.matches(&"some.pre2.test".into()).0);
+        assert!(!p.matches(&"some.pre2.test.2".into()).0);
+    }
+
+    #[test]
+    fn utest_empty_allow_path_pattern() {
+        let p = AllowPathPattern::from("");
+
+        assert!(p.matches(&"".into()).0);
+        assert!(p.matches(&"some.pre".into()).0);
+    }
+
+    #[test]
+    fn utest_deny_path_pattern() {
+        let p = DenyPathPattern::from("some.pre.fix");
+
+        assert!(p.matches(&"".into()).0);
+        assert!(p.matches(&"some.pre".into()).0);
+        assert!(p.matches(&"some.pre.fix".into()).0);
+        assert!(p.matches(&"some.pre.fix.test".into()).0);
+        assert!(!p.matches(&"some2.pre".into()).0);
+        assert!(!p.matches(&"some2.pre.fix".into()).0);
+        assert!(!p.matches(&"some.pre.fix2".into()).0);
+        assert!(!p.matches(&"some.pre.fix2.test".into()).0);
+    }
+
+    #[test]
+    fn utest_deny_path_pattern_with_wildcard() {
+        let p = DenyPathPattern::from("some.*.fix");
+
+        assert!(p.matches(&"".into()).0);
+        assert!(p.matches(&"some.pre".into()).0);
+        assert!(p.matches(&"some.pre.fix".into()).0);
+        assert!(p.matches(&"some.pre.fix.test".into()).0);
+        assert!(!p.matches(&"some2.pre".into()).0);
+        assert!(!p.matches(&"some2.pre.fix".into()).0);
+        assert!(!p.matches(&"some.pre.fix2".into()).0);
+        assert!(!p.matches(&"some.pre.fix2.test".into()).0);
+        assert!(p.matches(&"some.pre2".into()).0);
+        assert!(p.matches(&"some.pre2.fix".into()).0);
+        assert!(p.matches(&"some.pre2.fix.test".into()).0);
+        assert!(!p.matches(&"some2.pre2".into()).0);
+        assert!(!p.matches(&"some2.pre2.fix".into()).0);
+        assert!(!p.matches(&"some.pre2.fix2".into()).0);
+        assert!(!p.matches(&"some.pre2.fix2.test".into()).0);
+    }
+
+    #[test]
+    fn utest_empty_deny_path_pattern() {
+        let p = DenyPathPattern::from("");
+
+        assert!(p.matches(&"".into()).0);
+        assert!(p.matches(&"some.pre".into()).0);
+    }
+}
